Code cleanup and add more log.
